Movie Rating/requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
google-auth==1.30.0
h5py==3.2.1
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
MarkupSafe==2.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.24
mysqlclient==2.0.3
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.52
Pillow==8.2.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==3.7.7
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
soupsieve==2.2.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tflearn==0.5.0
tk==0.1.0
urllib3==1.26.4
Werkzeug==2.0.0
